requirements.txt
flask
imageio
numpy
requests
torch
scipy
pillow
pyyaml
matplotlib